Refactor: DebateRoom에서 커스텀 훅을 useMessage로 옮겨 가독성 향상
