+++ src/page/DebateRoom/DebateRoom/DebateRoom.jsx
@@ -6,48 +6,15 @@ import {MessageItem} from "../MessageItem/MessageItem";
 import {TopNavBar} from "../TopNavBar/TopNavBar";
 import styles from './DebateRoom.module.css';
 import {BottomNavBar} from "../BottomNavBar/BottomNavBar";
+import {useMessages} from "../utils/useMessages";
 
 export function DebateRoom() {
     const stompClient = useStomp();
     let { roomId } = useParams();
     const [yourMessage, setYourMessage] = useState('');
-    const [messages, setMessages] = useState({});
     const [messageThreads, setMessageThreads] = useState({});
-    const [agreeRatio, setAgreeRatio] = useState(0);
-    const [disagreeRatio, setDisagreeRatio] = useState(0);
+    const { messages, agreeRatio, disagreeRatio, handlePutPreference } = useMessages(roomId, stompClient);
 
-    useEffect(() => {
-        const fetchMessages = async () => {
-            const dataList = await fetchUtil(`/message/${roomId}`, {
-                method: 'GET'
-            });// 비동기 결과를 직접 상태에 할당
-
-            const messagesObject = dataList.reduce((acc, cur) => {
-                acc[cur.messageId] = cur;
-                return acc;
-            }, {});
-
-            setMessages(messagesObject);
-        };
-        fetchMessages();
-    }, [roomId]); // roomId를 의존성 배열에 추가
-
-    useEffect(() => {
-        if (!stompClient || stompClient.connected === false) {
-            console.log('Stomp client is not connected. Attempting to reconnect...');
-            stompClient.connect();
-        }
-        const url = `/sub/room/${roomId}`;
-        console.log(`subscribe room: ${roomId}`);
-        const subscription = stompClient.subscribe(url, function (chat) {
-            const message = JSON.parse(chat.body);
-            setMessages((prevMessages) => [...prevMessages, message]);
-        });
-
-        return () => {
-            subscription.unsubscribe();
-        };
-    }, [roomId, stompClient]);
 
     function sendMessage() {
         if (stompClient) {
@@ -65,38 +32,6 @@ export function DebateRoom() {
             ...prevThread,
             [messageId]: prevThreads.data
         }));
-    }
-    // likesNum을 업데이트하는 함수
-    const updateLikesNum = (messageId, newLikesNum) => {
-        setMessages((prevMessages) => ({
-            ...prevMessages,
-            [messageId]: { ...prevMessages[messageId], likesNum: newLikesNum }
-        }));
-    };
-
-    async function handlePutPreference(messageId, position){
-        const isAgree = position !== 'DISAGREE';
-        const dto = await fetchUtil(`/message/preference/${messageId}`, {
-            method: 'PUT',
-            body: {
-                isAgree: isAgree
-            }
-        });
-
-        setAgreeRatio(dto.ratio)
-        if(dto.ratio === 0  && disagreeRatio ===0){
-            setDisagreeRatio(dto.ratio)
-        }else{
-            setDisagreeRatio(100 - dto.ratio)
-        }
-
-        let newLikesNum = messages[messageId].likesNum;
-        if(dto.isIncrease){
-            newLikesNum += 1;
-        }else{
-            newLikesNum -= 1;
-        }
-        updateLikesNum(messageId, newLikesNum);
     }
     return (
         <div>
